@@ -8,44 +8,49 @@
     "start": "next start",
     "lint": "next lint",
     "prettier": "prettier --write src",
-    "prettier-check": "prettier --check src"
+    "prettier-check": "prettier --check src",
+    "stylelint": "stylelint **/*.css"
   },
   "dependencies": {
-    "@mantine/carousel": "^7.13.2",
-    "@mantine/core": "^7.13.2",
-    "@mantine/form": "^7.13.4",
-    "@mantine/hooks": "^7.13.2",
+    "@mantine/carousel": "^7.15.0",
+    "@mantine/core": "^7.15.0",
+    "@mantine/form": "^7.15.0",
+    "@mantine/hooks": "^7.15.0",
     "clsx": "^2.1.1",
-    "embla-carousel-react": "^7.0.0",
-    "i18next": "^23.16.0",
-    "next": "^14.2.15",
-    "next-i18next": "^15.3.1",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-i18next": "^15.0.3",
+    "embla-carousel-react": "^7.1.0",
+    "i18next": "^24.1.0",
+    "next": "^15.1.0",
+    "next-i18next": "^15.4.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-i18next": "^15.1.4",
     "telegram-escape": "^1.1.1",
-    "typescript": "^5.5.0"
+    "typescript": "^5.7.2"
   },
   "devDependencies": {
+    "@eslint/eslintrc": "^3.2.0",
     "@svgr/webpack": "^8.1.0",
-    "@trivago/prettier-plugin-sort-imports": "^4.3.0",
+    "@trivago/prettier-plugin-sort-imports": "^5.2.0",
     "@types/fluent-ffmpeg": "^2.1.27",
-    "@types/node": "^22.9.0",
-    "@types/react": "^18.3.1",
-    "@types/react-dom": "^18.3.1",
+    "@types/node": "^22.10.2",
+    "@types/react": "^19.0.1",
+    "@types/react-dom": "^19.0.2",
     "@types/youtube": "^0.1.0",
-    "eslint": "^8.57.0",
-    "eslint-config-next": "^14.2.15",
+    "eslint": "^9.16.0",
+    "eslint-config-next": "^15.1.0",
     "eslint-config-prettier": "^8.10.0",
     "eslint-plugin-prettier": "^5.2.1",
     "fluent-ffmpeg": "^2.1.3",
-    "postcss": "^8.4.47",
+    "postcss": "^8.4.49",
     "postcss-preset-mantine": "^1.17.0",
     "postcss-simple-vars": "^7.0.1",
-    "prettier": "^3.3.3",
+    "prettier": "^3.4.2",
     "sharp": "^0.33.5",
-    "stylelint": "^16.9.0",
+    "stylelint": "^16.11.0",
     "stylelint-config-clean-order": "^6.1.0",
     "stylelint-config-standard-scss": "^13.1.0"
+  }, 
+  "overrides": {
+    "react": "$react"
   }
 }
